Fixed interval selection for historycal location view.

+++ src/java/netxms-eclipse/OSM/src/org/netxms/ui/eclipse/osm/views/HistoryView.java
@@ -52,7 +52,7 @@ public class HistoryView extends ViewPart
 {
    public static final String ID = "org.netxms.ui.eclipse.osm.views.HistoryView"; //$NON-NLS-1$
 	public static final String JOB_FAMILY = "MapViewJob"; //$NON-NLS-1$
-	private static final int[] presetRanges = { 10, 30, 60, 120, 240, 720, 1440, 2880, 7200, 10080, 44640, 525600 };
+	private static final long[] presetRanges = { 10, 30, 60, 120, 240, 720, 1440, 2880, 7200, 10080, 44640, 525600 };
    private static final String[] presetNames = 
       { "10 minutes", "30 minutes", "1 hour", "2 hours", "4 hours", "12 hours", "Today",
         "Last 2 days", "Last 5 days", "This week", "This month","This Year" };
@@ -203,7 +203,7 @@ public class HistoryView extends ViewPart
             int result = dialog.open();
             if (result == Window.CANCEL)
                return;
-            map.changeTimePeriod((int)dialog.getTimeInMinutes());
+            map.changeTimePeriod(dialog.getTimeInMinutes());
             if(dialog.getTimeFrameType() == GraphSettings.TIME_FRAME_FIXED)
             {
                map.changeTimePeriod(dialog.getTimeFrom());
